+++ internal/config/settings.go
@@ -26,7 +26,6 @@ type Settings struct {
 	// g8e operator configuration - mirrors protocol/constants/env_vars.json "g8eo" section
 	OperatorEndpoint  string
 	OperatorSessionID string
-	InternalAuthToken string
 	PKIDir            string
 	SecretsDir        string
 	PubSubCACert      string
@@ -35,9 +34,6 @@ type Settings struct {
 	DataDir           string
 	IPService         string
 	IPResolver        string
-
-	// Insecure MCP
-	InsecureMcpToken string
 
 	// System / process environment
 	Shell       string
@@ -67,7 +63,6 @@ func LoadSettings() Settings {
 	return Settings{
 		OperatorEndpoint:  readVar(constants.EnvVar.OperatorEndpoint),
 		OperatorSessionID: readVar(constants.EnvVar.OperatorSessionID),
-		InternalAuthToken: readVar(constants.EnvVar.InternalAuthToken),
 		PKIDir:            readVar(constants.EnvVar.PKIDir),
 		SecretsDir:        readVar(constants.EnvVar.SecretsDir),
 		PubSubCACert:      readVar(constants.EnvVar.PubSubCACert),
@@ -76,7 +71,6 @@ func LoadSettings() Settings {
 		DataDir:           readVar(constants.EnvVar.DataDir),
 		IPService:         readVar(constants.EnvVar.IPService),
 		IPResolver:        readVar(constants.EnvVar.IPResolver),
-		InsecureMcpToken:  readVar(constants.EnvVar.InsecureMcpToken),
 		Shell:             readVar(constants.EnvVar.Shell),
 		Lang:              readVar(constants.EnvVar.Lang),
 		Term:              readVar(constants.EnvVar.Term),
