+++ internal/config/settings_test.go
@@ -47,14 +47,6 @@ func TestLoadSettings_IPFields(t *testing.T) {
 	assert.Equal(t, "1.1.1.1:80", s.IPResolver)
 }
 
-func TestLoadSettings_InsecureMcp(t *testing.T) {
-	t.Setenv(marshaler.EnvVar(constants.EnvVar.InsecureMcpToken), "oc-token-secret")
-
-	s := LoadSettings()
-
-	assert.Equal(t, "oc-token-secret", s.InsecureMcpToken)
-}
-
 func TestLoadSettings_SystemEnvVars(t *testing.T) {
 	t.Setenv(marshaler.EnvVar(constants.EnvVar.Shell), "/bin/zsh")
 	t.Setenv(marshaler.EnvVar(constants.EnvVar.Lang), "en_US.UTF-8")
@@ -108,7 +100,6 @@ func TestLoadSettings_EmptyWhenEnvNotSet(t *testing.T) {
 	t.Setenv(marshaler.EnvVar(constants.EnvVar.DataDir), "")
 	t.Setenv(marshaler.EnvVar(constants.EnvVar.IPService), "")
 	t.Setenv(marshaler.EnvVar(constants.EnvVar.IPResolver), "")
-	t.Setenv(marshaler.EnvVar(constants.EnvVar.InsecureMcpToken), "")
 
 	s := LoadSettings()
 
@@ -119,7 +110,6 @@ func TestLoadSettings_EmptyWhenEnvNotSet(t *testing.T) {
 	assert.Empty(t, s.DataDir)
 	assert.Empty(t, s.IPService)
 	assert.Empty(t, s.IPResolver)
-	assert.Empty(t, s.InsecureMcpToken)
 }
 
 func TestLoadSettings_AllFieldsPresent(t *testing.T) {
@@ -130,7 +120,6 @@ func TestLoadSettings_AllFieldsPresent(t *testing.T) {
 	t.Setenv(marshaler.EnvVar(constants.EnvVar.DataDir), "/data")
 	t.Setenv(marshaler.EnvVar(constants.EnvVar.IPService), "https://ip.svc")
 	t.Setenv(marshaler.EnvVar(constants.EnvVar.IPResolver), "8.8.8.8:80")
-	t.Setenv(marshaler.EnvVar(constants.EnvVar.InsecureMcpToken), "oc")
 	t.Setenv(marshaler.EnvVar(constants.EnvVar.Shell), "/bin/bash")
 	t.Setenv(marshaler.EnvVar(constants.EnvVar.Lang), "C")
 	t.Setenv(marshaler.EnvVar(constants.EnvVar.Term), "dumb")
@@ -148,7 +137,6 @@ func TestLoadSettings_AllFieldsPresent(t *testing.T) {
 	assert.Equal(t, "/data", s.DataDir)
 	assert.Equal(t, "https://ip.svc", s.IPService)
 	assert.Equal(t, "8.8.8.8:80", s.IPResolver)
-	assert.Equal(t, "oc", s.InsecureMcpToken)
 	assert.Equal(t, "/bin/bash", s.Shell)
 	assert.Equal(t, "C", s.Lang)
 	assert.Equal(t, "dumb", s.Term)
