+++ internal/services/gateway/gateway_http_test.go
@@ -23,21 +23,15 @@ import (
 	"net/http"
 	"net/http/httptest"
 	"net/url"
-	"os"
 	"strings"
 	"testing"
 	"time"
-
-	"github.com/gorilla/websocket"
 
 	"github.com/g8e-ai/g8e/internal/config"
 	"github.com/g8e-ai/g8e/internal/constants"
 	"github.com/g8e-ai/g8e/internal/marshaler"
 	"github.com/g8e-ai/g8e/internal/models"
-	"github.com/g8e-ai/g8e/internal/responder"
-	"github.com/g8e-ai/g8e/internal/services/keystore"
 	"github.com/g8e-ai/g8e/internal/services/mcp"
-	"github.com/g8e-ai/g8e/internal/testutil"
 	"github.com/g8e-ai/g8e/protocol"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -45,152 +39,64 @@ import (
 
 func setupTestHTTPHandler(t *testing.T) (*HTTPHandler, *config.Config) {
 	t.Helper()
-	cfg := testutil.NewTestConfig(t)
-	logger := testutil.NewTestLogger()
+	infra := setupTestInfrastructure(t, false)
 
-	dbDir := t.TempDir()
-	pkiDir := t.TempDir()
-	secretsDir := t.TempDir()
-	db, err := OpenGatewayDBService(dbDir, secretsDir, logger, true)
-	require.NoError(t, err)
-	t.Cleanup(func() { db.Close() })
-
-	// Remove the secrets directory written by InitAppSettings
-	os.RemoveAll(secretsDir)
-	os.MkdirAll(secretsDir, 0755)
-
-	pubsub := NewPubSubBroker(logger)
-	t.Cleanup(func() { pubsub.Close() })
-
-	// Initialize SecretManager with test backend for keystore operations
-	backend, err := keystore.NewTestBackend()
-	require.NoError(t, err)
-	ks, err := keystore.NewWithBackend(t.TempDir(), logger, backend)
-	require.NoError(t, err)
-	require.NoError(t, ks.Initialize())
-	require.NoError(t, ks.EnsurePermissions())
-	sm := &SecretManager{
-		db:         db.db,
-		secretsDir: t.TempDir(),
-		logger:     logger,
-		keystore:   ks,
-	}
-
-	pki := newPKIAuthority(dbDir, pkiDir, db, sm, logger)
-	err = pki.EnsurePKI(nil)
-	require.NoError(t, err)
-
-	userSvc := NewUserService(db, logger)
-	personaSvc := NewPersonaService(db, logger)
-	resp := responder.New(logger)
-	auth := NewAuthService(db, pki, logger, userSvc, personaSvc, resp, secretsDir, nil, "")
-	sessionSvc := NewSessionService(db, logger)
-	reg := NewRegistrationService(db, pki, logger, userSvc, sessionSvc, &cfg.Gateway)
-	passkey, _ := NewPasskeyService(db, logger, &PasskeyConfig{RpID: "localhost", RpName: "g8e"})
 	mcpGateway := mcp.NewGatewayService(mcp.Dependencies{
-		Logger:          logger,
-		Responder:       resp,
-		SuspendedStore:  db,
-		MaxPayloadBytes: cfg.Gateway.MaxPayloadBytes,
+		Logger:          infra.Logger,
+		Responder:       infra.Responder,
+		SuspendedStore:  infra.DB,
+		MaxPayloadBytes: infra.Cfg.Gateway.MaxPayloadBytes,
 	})
 	h := newHTTPHandler(HTTPHandlerDependencies{
-		Cfg:               cfg,
-		Logger:            logger,
-		DB:                db,
-		Pubsub:            pubsub,
-		Auth:              auth,
-		PKI:               pki,
-		SessionSvc:        sessionSvc,
-		Reg:               reg,
-		Passkey:           passkey,
-		UserSvc:           userSvc,
-		Responder:         resp,
+		Cfg:               infra.Cfg,
+		Logger:            infra.Logger,
+		DB:                infra.DB,
+		Pubsub:            infra.Pubsub,
+		Auth:              infra.Auth,
+		PKI:               infra.PKI,
+		SessionSvc:        infra.SessionSvc,
+		Reg:               infra.Reg,
+		Passkey:           infra.Passkey,
+		UserSvc:           infra.UserSvc,
+		Responder:         infra.Responder,
 		MCPGateway:        mcpGateway,
 		IsReady:           func() bool { return true },
 		IsGovernanceReady: func() bool { return true },
 	})
-	return h, cfg
+	return h, infra.Cfg
 }
 
 func setupTestGatewayService(t *testing.T) (*GatewayService, *config.Config) {
 	t.Helper()
-	cfg := testutil.NewTestConfig(t)
-	logger := testutil.NewTestLogger()
+	infra := setupTestInfrastructure(t, true)
 
-	// Create a real DB service for the tests to use
-	dbDir := t.TempDir()
-	pkiDir := t.TempDir()
-	secretsDir := t.TempDir()
-	db, err := OpenGatewayDBService(dbDir, secretsDir, logger, true)
-	require.NoError(t, err)
-	t.Cleanup(func() { db.Close() })
-
-	// Remove the secrets directory written by InitAppSettings
-	os.RemoveAll(secretsDir)
-	os.MkdirAll(secretsDir, 0755)
-
-	pubsub := NewPubSubBroker(logger)
-	t.Cleanup(func() { pubsub.Close() })
-
-	// Reset test storage before creating keystore (similar to keystore_test.go pattern)
-	keystore.ResetTestStorage()
-
-	// Initialize SecretManager with test backend for complete test environment
-	// Use separate temp dir for keystore to avoid conflicts with secretsDir
-	keystoreDir := t.TempDir()
-	backend, err := keystore.NewTestBackend()
-	require.NoError(t, err)
-	ks, err := keystore.NewWithBackend(keystoreDir, logger, backend)
-	require.NoError(t, err)
-	require.NoError(t, ks.Initialize())
-	require.NoError(t, ks.EnsurePermissions())
-	sm := &SecretManager{
-		db:         db.db,
-		secretsDir: secretsDir,
-		logger:     logger,
-		keystore:   ks,
-	}
-
-	pki := newPKIAuthority(dbDir, pkiDir, db, sm, logger)
-	err = pki.EnsurePKI(nil)
-	require.NoError(t, err)
-
-	// Build all services with the initialized PKI
-	userSvc := NewUserService(db, logger)
-	personaSvc := NewPersonaService(db, logger)
-	resp := responder.New(logger)
-	auth := NewAuthService(db, pki, logger, userSvc, personaSvc, resp, secretsDir, nil, "")
-	sessionSvc := NewSessionService(db, logger)
-	reg := NewRegistrationService(db, pki, logger, userSvc, sessionSvc, &cfg.Gateway)
-	passkey, _ := NewPasskeyService(db, logger, &PasskeyConfig{RpID: "localhost", RpName: "g8e"})
 	mcpGateway := mcp.NewGatewayService(mcp.Dependencies{
-		Logger:          logger,
-		Responder:       resp,
-		SuspendedStore:  db,
-		MaxPayloadBytes: cfg.Gateway.MaxPayloadBytes,
+		Logger:          infra.Logger,
+		Responder:       infra.Responder,
+		SuspendedStore:  infra.DB,
+		MaxPayloadBytes: infra.Cfg.Gateway.MaxPayloadBytes,
 	})
 
-	cfg.Gateway.BootstrapPort = constants.Ports.OperatorBootstrapHttps
+	infra.Cfg.Gateway.BootstrapPort = constants.Ports.OperatorBootstrapHttps
 
-	// Build GatewayService directly with initialized components
 	ls := &GatewayService{
-		cfg:        cfg,
-		logger:     logger,
-		db:         db,
-		pubsub:     pubsub,
-		auth:       auth,
-		pki:        pki,
-		reg:        reg,
-		passkey:    passkey,
-		userSvc:    userSvc,
-		sessionSvc: sessionSvc,
+		cfg:        infra.Cfg,
+		logger:     infra.Logger,
+		db:         infra.DB,
+		pubsub:     infra.Pubsub,
+		auth:       infra.Auth,
+		pki:        infra.PKI,
+		reg:        infra.Reg,
+		passkey:    infra.Passkey,
+		userSvc:    infra.UserSvc,
+		sessionSvc: infra.SessionSvc,
 		mcpGateway: mcpGateway,
-		responder:  resp,
+		responder:  infra.Responder,
 	}
 
 	require.NoError(t, ls.initHandlersAndServers())
 
-	return ls, cfg
+	return ls, infra.Cfg
 }
 
 func TestReadBody(t *testing.T) {
@@ -397,177 +303,6 @@ func TestHandleHealth(t *testing.T) {
 		err = json.Unmarshal(rr.Body.Bytes(), &resp)
 		require.NoError(t, err)
 		assert.Equal(t, constants.GatewayModeStatusOK, resp.Status)
-	})
-}
-
-func TestHandleDB(t *testing.T) {
-	h, _ := setupTestHTTPHandler(t)
-
-	t.Run("BadRequest - no collection", func(t *testing.T) {
-		t.Parallel()
-		req := httptest.NewRequest(http.MethodGet, "/db/", nil)
-		rr := httptest.NewRecorder()
-		h.handleDB(rr, req)
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-	})
-
-	t.Run("BadRequest - no ID", func(t *testing.T) {
-		t.Parallel()
-		req := httptest.NewRequest(http.MethodGet, "/db/users/", nil)
-		rr := httptest.NewRecorder()
-		h.handleDB(rr, req)
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-	})
-
-	t.Run("PUT and GET", func(t *testing.T) {
-		data := map[string]string{"name": "alice"}
-		reqPut := httptest.NewRequest(http.MethodPut, "/db/settings/u1", bytes.NewReader(mustDocJSON(t, data)))
-		rrPut := httptest.NewRecorder()
-		h.handleDB(rrPut, reqPut)
-		assert.Equal(t, http.StatusOK, rrPut.Code)
-
-		reqGet := httptest.NewRequest(http.MethodGet, "/db/settings/u1", nil)
-		rrGet := httptest.NewRecorder()
-		h.handleDB(rrGet, reqGet)
-		assert.Equal(t, http.StatusOK, rrGet.Code)
-
-		var doc map[string]interface{}
-		err := json.Unmarshal(rrGet.Body.Bytes(), &doc)
-		require.NoError(t, err)
-		assert.Equal(t, "alice", doc["name"])
-	})
-
-	t.Run("PATCH", func(t *testing.T) {
-		patch := map[string]string{"role": "admin"}
-		reqPatch := httptest.NewRequest(http.MethodPatch, "/db/settings/u1", bytes.NewReader(mustDocJSON(t, patch)))
-		rrPatch := httptest.NewRecorder()
-		h.handleDB(rrPatch, reqPatch)
-		assert.Equal(t, http.StatusOK, rrPatch.Code)
-
-		reqGet := httptest.NewRequest(http.MethodGet, "/db/settings/u1", nil)
-		rrGet := httptest.NewRecorder()
-		h.handleDB(rrGet, reqGet)
-		var doc map[string]interface{}
-		json.Unmarshal(rrGet.Body.Bytes(), &doc)
-		assert.Equal(t, "alice", doc["name"])
-		assert.Equal(t, "admin", doc["role"])
-	})
-
-	t.Run("DELETE", func(t *testing.T) {
-		reqDel := httptest.NewRequest(http.MethodDelete, "/db/settings/u1", nil)
-		rrDel := httptest.NewRecorder()
-		h.handleDB(rrDel, reqDel)
-		assert.Equal(t, http.StatusOK, rrDel.Code)
-
-		reqGet := httptest.NewRequest(http.MethodGet, "/db/settings/u1", nil)
-		rrGet := httptest.NewRecorder()
-		h.handleDB(rrGet, reqGet)
-		assert.Equal(t, http.StatusNotFound, rrGet.Code)
-	})
-
-	t.Run("Query", func(t *testing.T) {
-		h.db.DocSet("items", "i1", mustDocJSON(t, map[string]int{"val": 10}))
-		h.db.DocSet("items", "i2", mustDocJSON(t, map[string]int{"val": 20}))
-
-		query := models.DocQueryRequest{
-			Limit: 1,
-		}
-		body, _ := json.Marshal(query)
-		reqQuery := httptest.NewRequest(http.MethodPost, "/db/items/_query", bytes.NewReader(body))
-		rrQuery := httptest.NewRecorder()
-		h.handleDB(rrQuery, reqQuery)
-		assert.Equal(t, http.StatusOK, rrQuery.Code)
-
-		var results []map[string]interface{}
-		err := json.Unmarshal(rrQuery.Body.Bytes(), &results)
-		require.NoError(t, err)
-		assert.Len(t, results, 1)
-	})
-
-	t.Run("Invalid JSON", func(t *testing.T) {
-		t.Parallel()
-		req := httptest.NewRequest(http.MethodPut, "/db/settings/u1", strings.NewReader("{invalid-json}"))
-		rr := httptest.NewRecorder()
-		h.handleDB(rr, req)
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.JSONEq(t, `{"error":"invalid JSON body"}`, rr.Body.String())
-	})
-
-	t.Run("PATCH not found", func(t *testing.T) {
-		t.Parallel()
-		req := httptest.NewRequest(http.MethodPatch, "/db/settings/nonexistent", bytes.NewReader(mustDocJSON(t, map[string]string{"foo": "bar"})))
-		rr := httptest.NewRecorder()
-		h.handleDB(rr, req)
-		assert.Equal(t, http.StatusNotFound, rr.Code)
-	})
-
-	t.Run("DELETE not found", func(t *testing.T) {
-		t.Parallel()
-		req := httptest.NewRequest(http.MethodDelete, "/db/users/nonexistent", nil)
-		rr := httptest.NewRecorder()
-		h.handleDB(rr, req)
-		assert.Equal(t, http.StatusNotFound, rr.Code)
-	})
-
-	t.Run("Method Not Allowed", func(t *testing.T) {
-		t.Parallel()
-		req := httptest.NewRequest(http.MethodPost, "/db/users/u1", nil)
-		rr := httptest.NewRecorder()
-		h.handleDB(rr, req)
-		assert.Equal(t, http.StatusMethodNotAllowed, rr.Code)
-	})
-
-	t.Run("Non-bootstrap mutations redirect to governance envelope", func(t *testing.T) {
-		t.Parallel()
-		tests := []struct {
-			method string
-			body   []byte
-		}{
-			{method: http.MethodPut, body: mustDocJSON(t, map[string]string{"name": "alice"})},
-			{method: http.MethodPatch, body: mustDocJSON(t, map[string]string{"role": "admin"})},
-			{method: http.MethodDelete},
-		}
-
-		for _, tt := range tests {
-			req := httptest.NewRequest(tt.method, "/db/items/i1", bytes.NewReader(tt.body))
-			rr := httptest.NewRecorder()
-			h.handleDB(rr, req)
-			assert.Equal(t, http.StatusConflict, rr.Code, "method=%s", tt.method)
-			assert.JSONEq(t, `{"error":"submit via POST /api/governance/envelope"}`, rr.Body.String())
-		}
-	})
-
-	t.Run("Query validation", func(t *testing.T) {
-		t.Parallel()
-		req := httptest.NewRequest(http.MethodPost, "/db/items/_query", strings.NewReader("{invalid}"))
-		rr := httptest.NewRecorder()
-		h.handleDBQuery(rr, req, "items")
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-	})
-
-	t.Run("SSE Events count", func(t *testing.T) {
-		t.Parallel()
-		h.db.SSEEventsAppend(SSERoute{WebSessionID: "s1"}, "T", "{}")
-		req := httptest.NewRequest(http.MethodGet, "/db/_sse_events/count", nil)
-		rr := httptest.NewRecorder()
-		h.handleSSEEvents(rr, req, "count")
-		assert.Equal(t, http.StatusOK, rr.Code)
-	})
-
-	t.Run("SSE Events wipe", func(t *testing.T) {
-		t.Parallel()
-		req := httptest.NewRequest(http.MethodDelete, "/db/_sse_events", nil)
-		rr := httptest.NewRecorder()
-		h.handleSSEEvents(rr, req, "")
-		assert.Equal(t, http.StatusOK, rr.Code)
-	})
-
-	t.Run("SSE Events invalid", func(t *testing.T) {
-		t.Parallel()
-		req := httptest.NewRequest(http.MethodGet, "/db/_sse_events/invalid", nil)
-		rr := httptest.NewRecorder()
-		h.handleSSEEvents(rr, req, "invalid")
-		assert.Equal(t, http.StatusMethodNotAllowed, rr.Code)
 	})
 }
 
@@ -848,501 +583,6 @@ func TestInternalSSEBridge(t *testing.T) {
 		assert.Equal(t, http.StatusOK, rr.Code)
 		assert.Contains(t, rr.Header().Get("Content-Type"), "text/event-stream")
 		assert.Contains(t, rr.Body.String(), "event: stream-init")
-	})
-}
-
-func TestHandleKV(t *testing.T) {
-	h, _ := setupTestHTTPHandler(t)
-
-	t.Run("PUT and GET", func(t *testing.T) {
-		reqPut := httptest.NewRequest(http.MethodPut, "/kv/k1", bytes.NewReader(mustDocJSON(t, models.KVSetRequest{Value: "g8e"})))
-		rrPut := httptest.NewRecorder()
-		h.handleKV(rrPut, reqPut)
-		assert.Equal(t, http.StatusOK, rrPut.Code)
-
-		reqGet := httptest.NewRequest(http.MethodGet, "/kv/k1", nil)
-		rrGet := httptest.NewRecorder()
-		h.handleKV(rrGet, reqGet)
-		assert.Equal(t, http.StatusOK, rrGet.Code)
-		assert.Contains(t, rrGet.Body.String(), `"value":"g8e"`)
-	})
-
-	t.Run("TTL and Expire", func(t *testing.T) {
-		reqTtl := httptest.NewRequest(http.MethodGet, "/kv/k1/_ttl", nil)
-		rrTtl := httptest.NewRecorder()
-		h.handleKV(rrTtl, reqTtl)
-		assert.Equal(t, http.StatusOK, rrTtl.Code)
-
-		reqExp := httptest.NewRequest(http.MethodPut, "/kv/k1/_expire", bytes.NewReader(mustDocJSON(t, models.KVExpireRequest{TTL: 100})))
-		rrExp := httptest.NewRecorder()
-		h.handleKV(rrExp, reqExp)
-		assert.Equal(t, http.StatusOK, rrExp.Code)
-	})
-
-	t.Run("Scan and DeletePattern", func(t *testing.T) {
-		h.db.KVSet("pref:1", "a", 0)
-		h.db.KVSet("pref:2", "b", 0)
-
-		reqScan := httptest.NewRequest(http.MethodPost, "/kv/_scan", bytes.NewReader(mustDocJSON(t, models.KVPatternRequest{Pattern: "pref:*"})))
-		rrScan := httptest.NewRecorder()
-		h.handleKV(rrScan, reqScan)
-		assert.Equal(t, http.StatusOK, rrScan.Code)
-		assert.Contains(t, rrScan.Body.String(), "pref:1")
-
-		reqDel := httptest.NewRequest(http.MethodPost, "/kv/_delete_pattern", bytes.NewReader(mustDocJSON(t, models.KVPatternRequest{Pattern: "pref:*"})))
-		rrDel := httptest.NewRecorder()
-		h.handleKV(rrDel, reqDel)
-		assert.Equal(t, http.StatusOK, rrDel.Code)
-		assert.Contains(t, rrDel.Body.String(), `"deleted":2`)
-	})
-
-	t.Run("Invalid JSON", func(t *testing.T) {
-		t.Parallel()
-		req := httptest.NewRequest(http.MethodPut, "/kv/k1", strings.NewReader("{invalid-json}"))
-		rr := httptest.NewRecorder()
-		h.handleKV(rr, req)
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-	})
-
-	t.Run("TTL required for expire", func(t *testing.T) {
-		t.Parallel()
-		req := httptest.NewRequest(http.MethodPut, "/kv/k1/_expire", strings.NewReader(`{"ttl":0}`))
-		rr := httptest.NewRecorder()
-		h.handleKV(rr, req)
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-	})
-
-	t.Run("KV Keys", func(t *testing.T) {
-		t.Parallel()
-		h.db.KVSet("key1", "val1", 0)
-		req := httptest.NewRequest(http.MethodPost, "/kv/_keys", strings.NewReader(`{"pattern":"key*"}`))
-		rr := httptest.NewRecorder()
-		h.handleKVKeys(rr, req)
-		assert.Equal(t, http.StatusOK, rr.Code)
-		assert.Contains(t, rr.Body.String(), "key1")
-	})
-
-	t.Run("KV Keys Invalid JSON", func(t *testing.T) {
-		t.Parallel()
-		req := httptest.NewRequest(http.MethodPost, "/kv/_keys", strings.NewReader(`{invalid}`))
-		rr := httptest.NewRecorder()
-		h.handleKVKeys(rr, req)
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-	})
-
-	t.Run("KV Scan Invalid JSON", func(t *testing.T) {
-		t.Parallel()
-		req := httptest.NewRequest(http.MethodPost, "/kv/_scan", strings.NewReader(`{invalid}`))
-		rr := httptest.NewRecorder()
-		h.handleKVScan(rr, req)
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-	})
-
-	t.Run("KV Delete Pattern Missing Pattern", func(t *testing.T) {
-		t.Parallel()
-		req := httptest.NewRequest(http.MethodPost, "/kv/_delete_pattern", strings.NewReader(`{}`))
-		rr := httptest.NewRecorder()
-		h.handleKVDeletePattern(rr, req)
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-	})
-
-	t.Run("KV Delete Pattern Invalid JSON", func(t *testing.T) {
-		t.Parallel()
-		req := httptest.NewRequest(http.MethodPost, "/kv/_delete_pattern", strings.NewReader(`{invalid}`))
-		rr := httptest.NewRecorder()
-		h.handleKVDeletePattern(rr, req)
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-	})
-
-	t.Run("KV Method Not Allowed", func(t *testing.T) {
-		t.Parallel()
-		req := httptest.NewRequest(http.MethodPost, "/kv/key", nil)
-		rr := httptest.NewRecorder()
-		h.handleKV(rr, req)
-		assert.Equal(t, http.StatusMethodNotAllowed, rr.Code)
-	})
-}
-
-func TestHandleBlob(t *testing.T) {
-	t.Parallel()
-	h, _ := setupTestHTTPHandler(t)
-
-	t.Run("PUT and GET", func(t *testing.T) {
-		t.Parallel()
-		content := []byte("blob-data")
-		reqPut := httptest.NewRequest(http.MethodPut, "/blob/ns1/b1", bytes.NewReader(content))
-		reqPut.Header.Set("Content-Type", "text/plain")
-		rrPut := httptest.NewRecorder()
-		h.handleBlob(rrPut, reqPut)
-		assert.Equal(t, http.StatusOK, rrPut.Code)
-
-		reqGet := httptest.NewRequest(http.MethodGet, "/blob/ns1/b1", nil)
-		rrGet := httptest.NewRecorder()
-		h.handleBlob(rrGet, reqGet)
-		assert.Equal(t, http.StatusOK, rrGet.Code)
-		assert.Equal(t, content, rrGet.Body.Bytes())
-		assert.Equal(t, "text/plain", rrGet.Header().Get("Content-Type"))
-	})
-
-	t.Run("Metadata", func(t *testing.T) {
-		t.Parallel()
-		// Create a blob first
-		content := []byte("blob-data")
-		reqPut := httptest.NewRequest(http.MethodPut, "/blob/ns2/b2", bytes.NewReader(content))
-		reqPut.Header.Set("Content-Type", "text/plain")
-		rrPut := httptest.NewRecorder()
-		h.handleBlob(rrPut, reqPut)
-		assert.Equal(t, http.StatusOK, rrPut.Code)
-
-		// Then get metadata
-		reqMeta := httptest.NewRequest(http.MethodGet, "/blob/ns2/b2/meta", nil)
-		rrMeta := httptest.NewRecorder()
-		h.handleBlob(rrMeta, reqMeta)
-		assert.Equal(t, http.StatusOK, rrMeta.Code)
-		assert.Contains(t, rrMeta.Body.String(), `"id":"b2"`)
-	})
-
-	t.Run("Too Large", func(t *testing.T) {
-		t.Parallel()
-		largeBody := make([]byte, maxBlobBodySize+1)
-		req := httptest.NewRequest(http.MethodPut, "/blob/ns1/large", bytes.NewReader(largeBody))
-		req.Header.Set("Content-Type", "application/octet-stream")
-		rr := httptest.NewRecorder()
-		h.handleBlob(rr, req)
-		assert.Equal(t, http.StatusRequestEntityTooLarge, rr.Code)
-	})
-
-	t.Run("Missing Content-Type", func(t *testing.T) {
-		t.Parallel()
-		req := httptest.NewRequest(http.MethodPut, "/blob/ns1/b1", strings.NewReader("data"))
-		rr := httptest.NewRecorder()
-		h.handleBlob(rr, req)
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.JSONEq(t, `{"error":"Content-Type header required"}`, rr.Body.String())
-	})
-
-	t.Run("Invalid namespace", func(t *testing.T) {
-		t.Parallel()
-		req := httptest.NewRequest(http.MethodDelete, "/blob/../invalid", nil)
-		rr := httptest.NewRecorder()
-		h.handleBlob(rr, req)
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-	})
-
-	t.Run("Namespace delete not allowed method", func(t *testing.T) {
-		t.Parallel()
-		req := httptest.NewRequest(http.MethodPost, "/blob/ns1", nil)
-		rr := httptest.NewRecorder()
-		h.handleBlob(rr, req)
-		assert.Equal(t, http.StatusMethodNotAllowed, rr.Code)
-	})
-
-	t.Run("Blob id invalid", func(t *testing.T) {
-		t.Parallel()
-		req := httptest.NewRequest(http.MethodGet, "/blob/ns1/..", nil)
-		rr := httptest.NewRecorder()
-		h.handleBlob(rr, req)
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-	})
-
-	t.Run("Blob meta not found", func(t *testing.T) {
-		t.Parallel()
-		req := httptest.NewRequest(http.MethodGet, "/blob/ns1/nonexistent/meta", nil)
-		rr := httptest.NewRecorder()
-		h.handleBlob(rr, req)
-		assert.Equal(t, http.StatusNotFound, rr.Code)
-	})
-
-	t.Run("Blob get not found", func(t *testing.T) {
-		t.Parallel()
-		req := httptest.NewRequest(http.MethodGet, "/blob/ns1/nonexistent", nil)
-		rr := httptest.NewRecorder()
-		h.handleBlob(rr, req)
-		assert.Equal(t, http.StatusNotFound, rr.Code)
-	})
-
-	t.Run("Blob PUT empty body", func(t *testing.T) {
-		t.Parallel()
-		req := httptest.NewRequest(http.MethodPut, "/blob/ns1/empty", strings.NewReader(""))
-		req.Header.Set("Content-Type", "text/plain")
-		rr := httptest.NewRecorder()
-		h.handleBlob(rr, req)
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-		assert.JSONEq(t, `{"error":"body must not be empty"}`, rr.Body.String())
-	})
-
-	t.Run("Blob PUT read error", func(t *testing.T) {
-		t.Parallel()
-		// Create a request with a body that returns an error
-		req := httptest.NewRequest(http.MethodPut, "/blob/ns1/error", &errorReader{})
-		req.Header.Set("Content-Type", "text/plain")
-		rr := httptest.NewRecorder()
-		h.handleBlob(rr, req)
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-	})
-
-	t.Run("X-Blob-TTL valid", func(t *testing.T) {
-		t.Parallel()
-		req := httptest.NewRequest(http.MethodPut, "/blob/ns1/ttl-test", strings.NewReader("data"))
-		req.Header.Set("Content-Type", "text/plain")
-		req.Header.Set("X-Blob-TTL", "3600")
-		rr := httptest.NewRecorder()
-		h.handleBlob(rr, req)
-		assert.Equal(t, http.StatusOK, rr.Code)
-	})
-
-	t.Run("X-Blob-TTL invalid", func(t *testing.T) {
-		t.Parallel()
-		req := httptest.NewRequest(http.MethodPut, "/blob/ns1/ttl-fail", strings.NewReader("data"))
-		req.Header.Set("Content-Type", "text/plain")
-		req.Header.Set("X-Blob-TTL", "-1")
-		rr := httptest.NewRecorder()
-		h.handleBlob(rr, req)
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-	})
-}
-
-func TestWebSocketAuthIntegration(t *testing.T) {
-	t.Parallel()
-	h, _ := setupTestHTTPHandler(t)
-
-	server := httptest.NewServer(h.buildRouter())
-	defer server.Close()
-
-	wsURL := "ws" + strings.TrimPrefix(server.URL, "http") + "/ws/pubsub"
-
-	t.Run("Missing token", func(t *testing.T) {
-		ws, resp, err := websocket.DefaultDialer.Dial(wsURL, nil)
-		assert.Error(t, err)
-		if resp != nil && resp.Body != nil {
-			resp.Body.Close()
-		}
-		if ws != nil {
-			ws.Close()
-		}
-		require.NotNil(t, resp, "response should not be nil even on dial error")
-		assert.Equal(t, http.StatusUnauthorized, resp.StatusCode)
-	})
-}
-
-func TestHandlePubSubPublish(t *testing.T) {
-	t.Parallel()
-	h, _ := setupTestHTTPHandler(t)
-
-	t.Run("Publish valid", func(t *testing.T) {
-		t.Parallel()
-		pubReq := models.PubSubPublishRequest{
-			Channel: constants.ResultsChannel("op-1", "session-1"),
-			Data:    mustDocJSON(t, map[string]string{"foo": "bar"}),
-		}
-		body, _ := json.Marshal(pubReq)
-		req := httptest.NewRequest(http.MethodPost, "/pubsub/publish", bytes.NewReader(body))
-		rr := httptest.NewRecorder()
-		h.handlePubSubPublish(rr, req)
-		assert.Equal(t, http.StatusOK, rr.Code)
-		assert.Contains(t, rr.Body.String(), `"receivers":0`)
-	})
-
-	t.Run("Method Not Allowed", func(t *testing.T) {
-		t.Parallel()
-		req := httptest.NewRequest(http.MethodGet, "/pubsub/publish", nil)
-		rr := httptest.NewRecorder()
-		h.handlePubSubPublish(rr, req)
-		assert.Equal(t, http.StatusMethodNotAllowed, rr.Code)
-	})
-
-	t.Run("Invalid JSON", func(t *testing.T) {
-		t.Parallel()
-		req := httptest.NewRequest(http.MethodPost, "/pubsub/publish", strings.NewReader("{invalid}"))
-		rr := httptest.NewRecorder()
-		h.handlePubSubPublish(rr, req)
-		assert.Equal(t, http.StatusBadRequest, rr.Code)
-	})
-
-	t.Run("Reject mutation channels", func(t *testing.T) {
-		t.Parallel()
-		for _, channel := range []string{constants.CmdChannel("op-1", "session-1"), "auditor:op-1:session-1"} {
-			pubReq := models.PubSubPublishRequest{
-				Channel: channel,
-				Data:    mustDocJSON(t, map[string]string{"foo": "bar"}),
-			}
-			body, _ := json.Marshal(pubReq)
-			req := httptest.NewRequest(http.MethodPost, "/pubsub/publish", bytes.NewReader(body))
-			rr := httptest.NewRecorder()
-			h.handlePubSubPublish(rr, req)
-			assert.Equal(t, http.StatusConflict, rr.Code, "channel=%s", channel)
-			assert.JSONEq(t, `{"error":"submit via POST /api/governance/envelope"}`, rr.Body.String())
-		}
-	})
-}
-
-func TestHandleBootstrap(t *testing.T) {
-	t.Parallel()
-	h, _ := setupTestHTTPHandler(t)
-
-	t.Run("Success - Bootstrap with CSR over loopback", func(t *testing.T) {
-		t.Parallel()
-		csr := generateTestCSR(t)
-		cliCsr := generateTestCSR(t)
-		body := map[string]string{
-			"name":               "Owner",
-			"csr_pem":            csr,
-			"cli_csr_pem":        cliCsr,
-			"system_fingerprint": "test-fp",
-		}
-		b, _ := json.Marshal(body)
-		req := httptest.NewRequest(http.MethodPost, "/api/auth/bootstrap", bytes.NewReader(b))
-		req.RemoteAddr = "127.0.0.1:12345"
-		rr := httptest.NewRecorder()
-
-		h.handleBootstrap(rr, req)
-
-		assert.Equal(t, http.StatusCreated, rr.Code)
-		var resp map[string]interface{}
-		err := json.Unmarshal(rr.Body.Bytes(), &resp)
-		require.NoError(t, err)
-		assert.True(t, resp["success"].(bool))
-		assert.NotEmpty(t, resp["operator_cert"])
-		assert.NotEmpty(t, resp["operator_cert_chain"])
-		assert.NotEmpty(t, resp["hub_trust_bundle"])
-		assert.NotEmpty(t, resp["operator_session_id"])
-		assert.NotEmpty(t, resp["cli_session_id"])
-		assert.NotEqual(t, resp["operator_session_id"], resp["cli_session_id"],
-			"cli_session_id MUST be a distinct identifier from operator_session_id")
-	})
-
-	t.Run("Failure - Non-loopback CSR request rejected", func(t *testing.T) {
-		t.Parallel()
-		// Create a fresh handler to ensure no users exist
-		h2, _ := setupTestHTTPHandler(t)
-		csr := generateTestCSR(t)
-		cliCsr := generateTestCSR(t)
-		body := map[string]string{
-			"name":               "Owner",
-			"csr_pem":            csr,
-			"cli_csr_pem":        cliCsr,
-			"system_fingerprint": "test-fp",
-		}
-		b, _ := json.Marshal(body)
-		req := httptest.NewRequest(http.MethodPost, "/api/auth/bootstrap", bytes.NewReader(b))
-		req.RemoteAddr = "192.168.1.1:12345"
-		rr := httptest.NewRecorder()
-
-		h2.handleBootstrap(rr, req)
-
-		assert.Equal(t, http.StatusForbidden, rr.Code)
-		assert.JSONEq(t, `{"error":"CSR auto-issue only available over loopback"}`, rr.Body.String())
-	})
-
-	t.Run("Success - Rotation for existing bootstrap user", func(t *testing.T) {
-		t.Parallel()
-		hRot, _ := setupTestHTTPHandler(t)
-		// Create a bootstrap user first
-		user, err := hRot.userSvc.CreateBootstrapUser()
-		require.NoError(t, err)
-		require.NotNil(t, user)
-
-		csr := generateTestCSR(t)
-		cliCsr := generateTestCSR(t)
-		body := map[string]string{
-			"name":               "Owner",
-			"csr_pem":            csr,
-			"cli_csr_pem":        cliCsr,
-			"system_fingerprint": "rotated-fp",
-		}
-		b, _ := json.Marshal(body)
-		req := httptest.NewRequest(http.MethodPost, "/api/auth/bootstrap", bytes.NewReader(b))
-		req.RemoteAddr = "127.0.0.1:12345"
-		rr := httptest.NewRecorder()
-
-		hRot.handleBootstrap(rr, req)
-
-		assert.Equal(t, http.StatusCreated, rr.Code)
-		var resp map[string]interface{}
-		err = json.Unmarshal(rr.Body.Bytes(), &resp)
-		require.NoError(t, err)
-		assert.True(t, resp["success"].(bool))
-		assert.NotEmpty(t, resp["operator_cert"])
-	})
-
-	t.Run("Failure - Rotation fails for disabled bootstrap user", func(t *testing.T) {
-		t.Parallel()
-		h3, _ := setupTestHTTPHandler(t)
-		// 1. Bootstrap
-		user, _ := h3.userSvc.CreateBootstrapUser()
-		// 2. Disable
-		h3.userSvc.Disable(user.ID, "retired", "actor", "op")
-
-		// 3. Attempt rotation
-		csr := generateTestCSR(t)
-		cliCsr := generateTestCSR(t)
-		body := map[string]string{
-			"name":               "Owner",
-			"csr_pem":            csr,
-			"cli_csr_pem":        cliCsr,
-			"system_fingerprint": "fail-fp",
-		}
-		b, _ := json.Marshal(body)
-		req := httptest.NewRequest(http.MethodPost, "/api/auth/bootstrap", bytes.NewReader(b))
-		req.RemoteAddr = "127.0.0.1:12345"
-		rr := httptest.NewRecorder()
-
-		h3.handleBootstrap(rr, req)
-
-		assert.Equal(t, http.StatusConflict, rr.Code)
-		assert.JSONEq(t, `{"error":"bootstrap user is disabled, cannot rotate"}`, rr.Body.String())
-	})
-
-	t.Run("Failure - Rejects bootstrap if ANY other users exist", func(t *testing.T) {
-		t.Parallel()
-		h4, _ := setupTestHTTPHandler(t)
-		// Create a regular user first
-		h4.userSvc.CreateUser()
-
-		body := map[string]string{
-			"name": "Superadmin",
-		}
-		b, _ := json.Marshal(body)
-		req := httptest.NewRequest(http.MethodPost, "/api/auth/bootstrap", bytes.NewReader(b))
-		rr := httptest.NewRecorder()
-
-		h4.handleBootstrap(rr, req)
-
-		assert.Equal(t, http.StatusForbidden, rr.Code)
-		assert.JSONEq(t, `{"error":"bootstrap only available for initial setup"}`, rr.Body.String())
-	})
-}
-
-func TestHandleBootstrapStatus(t *testing.T) {
-	t.Run("Initially not bootstrapped", func(t *testing.T) {
-		t.Parallel()
-		h, _ := setupTestHTTPHandler(t)
-		req := httptest.NewRequest(http.MethodGet, "/api/auth/bootstrap/status", nil)
-		rr := httptest.NewRecorder()
-		h.handleBootstrapStatus(rr, req)
-
-		assert.Equal(t, http.StatusOK, rr.Code)
-		var resp map[string]interface{}
-		err := json.Unmarshal(rr.Body.Bytes(), &resp)
-		require.NoError(t, err)
-		assert.Equal(t, false, resp["bootstrapped"])
-	})
-
-	t.Run("Bootstrapped after creating a user", func(t *testing.T) {
-		t.Parallel()
-		h, _ := setupTestHTTPHandler(t)
-		_, err := h.userSvc.CreateUser()
-		require.NoError(t, err)
-
-		req := httptest.NewRequest(http.MethodGet, "/api/auth/bootstrap/status", nil)
-		rr := httptest.NewRecorder()
-		h.handleBootstrapStatus(rr, req)
-
-		assert.Equal(t, http.StatusOK, rr.Code)
-		var resp map[string]interface{}
-		err = json.Unmarshal(rr.Body.Bytes(), &resp)
-		require.NoError(t, err)
-		assert.Equal(t, true, resp["bootstrapped"])
 	})
 }
