+++ internal/services/gateway/pki_controller.go
@@ -1,0 +1,269 @@
+// Copyright (c) 2026 Lateralus Labs, LLC.
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package gateway
+
+import (
+	"crypto/sha256"
+	"encoding/hex"
+	"encoding/json"
+	"encoding/pem"
+	"io"
+	"log/slog"
+	"net/http"
+	"os"
+	"path/filepath"
+	"strings"
+	"time"
+
+	"github.com/g8e-ai/g8e/internal/config"
+	"github.com/g8e-ai/g8e/internal/constants"
+	"github.com/g8e-ai/g8e/internal/models"
+	"github.com/g8e-ai/g8e/internal/responder"
+)
+
+// PKIController handles PKI and certificate management endpoints.
+type PKIController struct {
+	cfg           *config.Config
+	logger        *slog.Logger
+	db            *GatewayDBService
+	pki           *PKIAuthority
+	appEnrollment *AppEnrollmentService
+	responder     *responder.Responder
+}
+
+func newPKIController(cfg *config.Config, logger *slog.Logger, db *GatewayDBService, pki *PKIAuthority, appEnrollment *AppEnrollmentService, responder *responder.Responder) *PKIController {
+	return &PKIController{
+		cfg:           cfg,
+		logger:        logger,
+		db:            db,
+		pki:           pki,
+		appEnrollment: appEnrollment,
+		responder:     responder,
+	}
+}
+
+func (c *PKIController) readBody(r *http.Request) ([]byte, error) {
+	r.Body = http.MaxBytesReader(nil, r.Body, c.cfg.Gateway.MaxPayloadBytes)
+	return io.ReadAll(r.Body)
+}
+
+func (c *PKIController) handlePKIHubBundle(w http.ResponseWriter, r *http.Request) {
+	if r.Method != http.MethodGet {
+		c.responder.Error(w, http.StatusMethodNotAllowed, "method not allowed")
+		return
+	}
+	pem, err := c.pki.HubTrustBundle()
+	if err != nil {
+		c.responder.Error(w, http.StatusInternalServerError, "failed to read hub bundle")
+		return
+	}
+	w.Header().Set(constants.HeaderContentType, "application/x-pem-file")
+	w.Header().Set("X-Content-Type-Options", "nosniff")
+	w.Header().Set("X-Frame-Options", "DENY")
+	w.WriteHeader(http.StatusOK)
+	_, _ = w.Write(pem)
+}
+
+func (c *PKIController) handlePKIFingerprint(w http.ResponseWriter, r *http.Request) {
+	if r.Method != http.MethodGet {
+		c.responder.Error(w, http.StatusMethodNotAllowed, "method not allowed")
+		return
+	}
+	pemData, err := os.ReadFile(filepath.Join(c.pki.PKIDir(), "root", "root_ca.crt"))
+	if err != nil {
+		c.responder.Error(w, http.StatusInternalServerError, "failed to read root CA")
+		return
+	}
+
+	block, _ := pem.Decode(pemData)
+	if block == nil {
+		c.responder.Error(w, http.StatusInternalServerError, "invalid root CA PEM")
+		return
+	}
+
+	hash := sha256.Sum256(block.Bytes)
+	fingerprint := hex.EncodeToString(hash[:])
+
+	c.responder.JSON(w, http.StatusOK, map[string]string{
+		"root_ca": "sha256:" + fingerprint,
+	})
+}
+
+func (c *PKIController) handlePKISignCSR(w http.ResponseWriter, r *http.Request) {
+	if r.Method != http.MethodPost {
+		c.responder.Error(w, http.StatusMethodNotAllowed, "method not allowed")
+		return
+	}
+
+	body, err := c.readBody(r)
+	if err != nil {
+		c.responder.Error(w, http.StatusBadRequest, "failed to read body")
+		return
+	}
+
+	var req struct {
+		CSR               string `json:"csr_pem"`
+		LeafType          string `json:"leaf_type"`
+		OrganizationID    string `json:"organization_id"`
+		OperatorID        string `json:"operator_id"`
+		UserID            string `json:"user_id"`
+		WorkloadSessionID string `json:"workload_session_id"`
+	}
+	if err := json.Unmarshal(body, &req); err != nil {
+		c.responder.Error(w, http.StatusBadRequest, "invalid JSON")
+		return
+	}
+
+	certPEM, chainPEM, err := c.pki.SignCSR(req.CSR, req.LeafType, req.OrganizationID, req.OperatorID, req.UserID, req.WorkloadSessionID)
+	if err != nil {
+		c.responder.Error(w, http.StatusInternalServerError, err.Error())
+		return
+	}
+
+	c.responder.JSON(w, http.StatusOK, map[string]string{
+		"certificate_pem":       certPEM,
+		"certificate_chain_pem": chainPEM,
+	})
+}
+
+func (c *PKIController) handlePKIRevoke(w http.ResponseWriter, r *http.Request) {
+	if r.Method != http.MethodPost {
+		c.responder.Error(w, http.StatusMethodNotAllowed, "method not allowed")
+		return
+	}
+
+	body, err := c.readBody(r)
+	if err != nil {
+		c.responder.Error(w, http.StatusBadRequest, "failed to read body")
+		return
+	}
+
+	var req struct {
+		Serial string `json:"serial"`
+		Reason string `json:"reason"`
+	}
+	if err := json.Unmarshal(body, &req); err != nil {
+		c.responder.Error(w, http.StatusBadRequest, "invalid JSON")
+		return
+	}
+
+	if req.Serial == "" {
+		c.responder.Error(w, http.StatusBadRequest, "serial required")
+		return
+	}
+
+	if err := c.pki.RevokeCertificate(req.Serial, req.Reason); err != nil {
+		c.responder.Error(w, http.StatusInternalServerError, err.Error())
+		return
+	}
+
+	c.responder.JSON(w, http.StatusOK, models.StatusResponse{Status: constants.GatewayModeStatusOK})
+}
+
+func (c *PKIController) handlePKIRevocationBundle(w http.ResponseWriter, r *http.Request) {
+	if r.Method != http.MethodGet {
+		c.responder.Error(w, http.StatusMethodNotAllowed, "method not allowed")
+		return
+	}
+
+	bundleJSON, signature, err := c.pki.GenerateRevocationBundle()
+	if err != nil {
+		c.responder.Error(w, http.StatusInternalServerError, err.Error())
+		return
+	}
+
+	c.responder.JSON(w, http.StatusOK, map[string]string{
+		"bundle_json": bundleJSON,
+		"signature":   signature,
+	})
+}
+
+func (c *PKIController) handleAppEnroll(w http.ResponseWriter, r *http.Request) {
+	if r.Method != http.MethodPost {
+		c.responder.Error(w, http.StatusMethodNotAllowed, "method not allowed")
+		return
+	}
+
+	if c.appEnrollment == nil {
+		c.responder.Error(w, http.StatusServiceUnavailable, "app enrollment service not available")
+		return
+	}
+
+	authHeader := r.Header.Get("Authorization")
+	if !strings.HasPrefix(authHeader, "Bearer ") {
+		c.responder.Error(w, http.StatusUnauthorized, "missing bearer token")
+		return
+	}
+
+	token := strings.TrimPrefix(authHeader, "Bearer ")
+	if !strings.HasPrefix(token, "dlk_") || len(token) < 20 {
+		c.responder.Error(w, http.StatusUnauthorized, "invalid device-link token format")
+		return
+	}
+
+	linkKey := "g8e:device-link:" + token
+	raw, found := c.db.KVGet(linkKey)
+	if !found {
+		c.responder.Error(w, http.StatusUnauthorized, "device-link token not found")
+		return
+	}
+
+	var linkData map[string]interface{}
+	if err := json.Unmarshal([]byte(raw), &linkData); err != nil {
+		c.responder.Error(w, http.StatusUnauthorized, "invalid device-link token data")
+		return
+	}
+
+	expiresAt, ok := linkData["expires_at"].(string)
+	if !ok {
+		c.responder.Error(w, http.StatusUnauthorized, "device-link token missing expiry")
+		return
+	}
+
+	expTime, err := time.Parse(time.RFC3339, expiresAt)
+	if err != nil {
+		c.responder.Error(w, http.StatusUnauthorized, "invalid device-link token expiry")
+		return
+	}
+
+	if expTime.Before(time.Now()) {
+		c.responder.Error(w, http.StatusUnauthorized, "device-link token expired")
+		return
+	}
+
+	body, err := c.readBody(r)
+	if err != nil {
+		c.responder.Error(w, http.StatusBadRequest, "failed to read body")
+		return
+	}
+
+	var req AppEnrollRequest
+	if err := json.Unmarshal(body, &req); err != nil {
+		c.responder.Error(w, http.StatusBadRequest, "invalid JSON")
+		return
+	}
+
+	resp, err := c.appEnrollment.EnrollApp(req)
+	if err != nil {
+		c.responder.Error(w, http.StatusInternalServerError, err.Error())
+		return
+	}
+
+	if !resp.Success {
+		c.responder.Error(w, http.StatusBadRequest, resp.Error)
+		return
+	}
+
+	c.responder.JSON(w, http.StatusCreated, resp)
+}
