+++ protocol/constants/env_vars.json
@@ -18,7 +18,6 @@
     "IPResolver": "G8E_IP_RESOLVER",
     "IPService": "G8E_IP_SERVICE",
     "InternalAPIKey": "G8E_INTERNAL_API_KEY",
-    "InternalAuthToken": "G8E_INTERNAL_AUTH_TOKEN",
     "LLMCommandGenAuditor": "G8E_LLM_COMMAND_GEN_AUDITOR",
     "LLMCommandGenEnabled": "G8E_LLM_COMMAND_GEN_ENABLED",
     "LLMCommandGenPasses": "G8E_LLM_COMMAND_GEN_PASSES",
@@ -32,7 +31,6 @@
     "LogLevel": "G8E_LOG_LEVEL",
     "LogMaxBackups": "G8E_LOG_MAX_BACKUPS",
     "LogName": "LOGNAME",
-    "InsecureMcpToken": "G8E_INSECURE_MCP_TOKEN",
     "OperatorBlobURL": "G8E_OPERATOR_BLOB_URL",
     "OperatorBootstrapHTTPSPort": "G8E_REMOTE_OPERATOR_BOOTSTRAP_HTTPS_PORT",
     "OperatorEndpoint": "G8E_OPERATOR_ENDPOINT",
@@ -57,7 +55,6 @@
     "RuntimeDir": "G8E_RUNTIME_DIR",
     "SSHAuthSock": "SSH_AUTH_SOCK",
     "SSHConfigPath": "G8E_SSH_CONFIG_PATH",
-    "SSLCertFile": "G8E_SSL_CERT_FILE",
     "SecretsDir": "G8E_SECRETS_DIR",
     "SessionEncryptionKey": "G8E_SESSION_ENCRYPTION_KEY",
     "SessionTTL": "G8E_SESSION_TTL",
